feat: create separate resource for build services
