+++ apps/api/src/apps/builds/build.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { MongooseModule } from '@nestjs/mongoose';
-import { UserModule } from '../../features/user/user.module';
+import { BuildModule } from '../../features/builds/build.module';
 
 @Module({
   imports: [
@@ -14,8 +14,8 @@ import { UserModule } from '../../features/user/user.module';
       inject: [ConfigService],
     }),
 
-    UserModule,
+    BuildModule,
   ],
   providers: [],
 })
-export class BuildModule {}
+export class BuildsModule {}
