+++ apps/api/src/features/builds/libs/tokens.ts
@@ -1,0 +1,3 @@
+export const Tokens = {
+  UserBuildRepository: Symbol('UserBuildRepository'),
+};
