+++ apps/api/src/features/user/libs/tokens.ts
@@ -1,4 +1,3 @@
 export const Tokens = {
   UserRepository: Symbol('UserRepository'),
-  UserBuildRepository: Symbol('UserBuildRepository'),
 };
