+++ apps/api/src/features/user/user.service.ts
@@ -2,18 +2,14 @@ import { Inject, Injectable } from '@nestjs/common';
 import R from 'ramda';
 import { Tokens } from './libs/tokens';
 import { UserRepository } from './repository/user.repository';
-import { UserBuildRepository } from './repository/user-build.repository';
-import { User, UserBuild } from '@pcp/types';
+import { User } from '@pcp/types';
 import { ObjectId } from '@pcp/object-id';
 
 @Injectable()
 export class UserService {
   constructor(
     @Inject(Tokens.UserRepository)
-    private userRepository: UserRepository,
-
-    @Inject(Tokens.UserBuildRepository)
-    private userBuildRepository: UserBuildRepository
+    private userRepository: UserRepository
   ) {}
 
   async createUser(data: R.Omit<User, 'dateTimeCreated' | 'dateTimeUpdated'>) {
@@ -45,15 +41,5 @@ export class UserService {
 
   async findUsers() {
     return this.userRepository.find({});
-  }
-
-  public async createBuild(
-    data: Omit<UserBuild, 'dateTimeCreated' | 'dateTimeUpdated'>
-  ) {
-    return this.userBuildRepository.create({
-      ...data,
-      dateTimeCreated: new Date(),
-      dateTimeUpdated: new Date(),
-    });
   }
 }
